whisper-local-server/requirements.txt
transformers>=4.35.0
torch>=2.0.0
torchaudio>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
accelerate>=0.25.0
